refactoring: remove unused miranum platform functions
