+++ examples/single-deployment-unit-example/src/main/resources/db/migration/postgresql/V08__revert_miranum_engine.sql
@@ -1,0 +1,7 @@
+drop index if exists IDX_DWF_PROCAUTH_USERID;
+drop index if exists IDX_DWF_FILTER_USERID;
+
+drop table if exists miranum_start_context;
+drop table if exists miranum_process_instance;
+drop table if exists miranum_process_instance_auth;
+drop table if exists miranum_filter;
