+++ examples/single-deployment-unit-example/src/main/resources/db/migration/postgresql/V09__revert_miranum_schema_registry.sql
@@ -1,0 +1,3 @@
+drop index if exists miranum_schema_registry_bundle_ref_tag_idx;
+drop index if exists miranum_schema_registry_bundle_ref_idx;
+drop table if exists miranum_schema_registry;
