+++ platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/api/MiranumProcessDefinition.java
@@ -1,4 +1,4 @@
-package io.miragon.miranum.platform.engine.domain.process;
+package io.miragon.miranum.platform.engine.api;
 
 import lombok.*;
