+++ platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/api/MiranumProcessDefinitionApi.java
@@ -1,0 +1,7 @@
+package io.miragon.miranum.platform.engine.api;
+
+public interface MiranumProcessDefinitionApi {
+
+    MiranumProcessDefinition getProcessDefinitionById(final String definitionId);
+
+}
