--- platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/application/port/in/process/ProcessInstanceQuery.java
@@ -1,14 +1,0 @@
-package io.miragon.miranum.platform.engine.application.port.in.process;
-
-import io.miragon.miranum.platform.engine.domain.process.MiranumProcessInstance;
-
-import java.util.List;
-
-/**
- * Service to interact with process instances.
- */
-public interface ProcessInstanceQuery {
-
-    List<MiranumProcessInstance> getProcessInstanceByUser(final String userId, String query);
-
-}
