--- platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/application/port/in/process/UpdateProcessInstanceDescriptionUseCase.java
@@ -1,7 +1,0 @@
-package io.miragon.miranum.platform.engine.application.port.in.process;
-
-public interface UpdateProcessInstanceDescriptionUseCase {
-
-    void updateDescription(String processInstanceId, String description);
-
-}
