--- platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/domain/file/PresignedUrlAction.java
@@ -1,8 +1,0 @@
-package io.miragon.miranum.platform.engine.domain.file;
-
-public enum PresignedUrlAction {
-    GET,
-    POST,
-    PUT,
-    DELETE
-}
