--- platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/domain/process/HistoryTask.java
@@ -1,33 +1,0 @@
-package io.miragon.miranum.platform.engine.domain.process;
-
-import lombok.*;
-
-import java.util.Date;
-
-/**
- * Task that is completed.
- *
- * @author externer.dl.horn
- */
-@Getter
-@Builder
-@ToString
-@EqualsAndHashCode
-@AllArgsConstructor
-public class HistoryTask {
-
-    /**
-     * Name of the task.
-     */
-    private final String id;
-
-    /**
-     * Name of the task.
-     */
-    private final String name;
-
-    /**
-     * Time when the task was completed.
-     */
-    private final Date endTime;
-}
