+++ platform/engine/task/pom.xml
@@ -57,16 +57,6 @@
             <artifactId>connect-task-api</artifactId>
             <version>${project.version}</version>
         </dependency>
-
-        <!-- Feign -->
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-starter-openfeign</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.github.openfeign</groupId>
-            <artifactId>feign-httpclient</artifactId>
-        </dependency>
     </dependencies>
 
 </project>
