+++ platform/engine/task/src/main/java/io/miragon/miranum/platform/tasklist/application/service/TaskInfoService.java
@@ -1,8 +1,8 @@
 package io.miragon.miranum.platform.tasklist.application.service;
 
 import io.holunda.camunda.bpm.data.factory.VariableFactory;
-import io.miragon.miranum.platform.engine.application.port.out.process.MiranumProcessDefinitionPort;
-import io.miragon.miranum.platform.engine.domain.process.MiranumProcessDefinition;
+import io.miragon.miranum.platform.engine.api.MiranumProcessDefinition;
+import io.miragon.miranum.platform.engine.api.MiranumProcessDefinitionApi;
 import io.miragon.miranum.platform.tasklist.TaskProperties;
 import io.miragon.miranum.platform.tasklist.application.port.in.TaskInfoUseCase;
 import io.miragon.miranum.platform.tasklist.application.port.out.task.TaskOutPort;
@@ -24,7 +24,7 @@ import static io.holunda.camunda.bpm.data.CamundaBpmData.stringVariable;
 public class TaskInfoService implements TaskInfoUseCase {
 
     private final TaskOutPort taskOutPort;
-    private final MiranumProcessDefinitionPort miranumProcessDefinitionPort;
+    private final MiranumProcessDefinitionApi miranumProcessDefinitionPort;
     private final TaskProperties properties;
 
     public static final String TASK_DESCRIPTION = "miranum_task_description";
