+++ platform/engine/task/src/test/java/io/miragon/miranum/platform/tasklist/application/service/TaskInfoServiceTest.java
@@ -1,7 +1,7 @@
 package io.miragon.miranum.platform.tasklist.application.service;
 
-import io.miragon.miranum.platform.engine.application.port.out.process.MiranumProcessDefinitionPort;
-import io.miragon.miranum.platform.engine.domain.process.MiranumProcessDefinition;
+import io.miragon.miranum.platform.engine.api.MiranumProcessDefinition;
+import io.miragon.miranum.platform.engine.api.MiranumProcessDefinitionApi;
 import io.miragon.miranum.platform.tasklist.TaskProperties;
 import io.miragon.miranum.platform.tasklist.application.port.in.TaskInfoUseCase;
 import io.miragon.miranum.platform.tasklist.application.port.out.task.TaskOutPort;
@@ -19,7 +19,7 @@ import static org.mockito.Mockito.*;
 class TaskInfoServiceTest {
 
     private final TaskOutPort taskOutPort = mock(TaskOutPort.class);
-    private final MiranumProcessDefinitionPort miranumProcessDefinitionPort = mock(MiranumProcessDefinitionPort.class);
+    private final MiranumProcessDefinitionApi miranumProcessDefinitionPort = mock(MiranumProcessDefinitionApi.class);
     private final TaskProperties taskProperties = new TaskProperties();
 
     private final TaskInfoUseCase taskInfoService = new TaskInfoService(taskOutPort, miranumProcessDefinitionPort, taskProperties);
